settings.json: Update `qtFolder` & remove `qtInstallations`
* Remove qtInstallations value from settings.json.
Those installation values are not needed to be written into
`settings.json`
* Add state manager
A state manager required to put value in `globalStorage`
* Change qtFolder scope from `machine-overridable` to `machine`
Having `qtFolder` config in both workspace and global creates conflict.
Because the Qt extension is still dependent to set `cmake.addtionalKits`
in a setting.json file.
* Remove `detectQtCMakeProject` command. That command is not used.
* Remove the file watcher for `qtfolder`. The current implementation
cannot detect anything in `qtfolder`. And when `**/*` pattern is used,
it fails with `Inotify limit reached` on Linux.
* Remove `filewatcher` for `qt-kits.json

Change-Id: I1a01f9e1d45538d40a331629c907ad530a8a0d37
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>