Add package-lock.json to the repo
This allows us to install a well-defined set of dependencies in CI, and
on local machines. We also now recommend to run "npm ci" instead of
"npm install", because the former respects the pinned dependencies.

The project maintainer may at some point run "npm install" to update the
dependencies in package-lock.json and commit the updated file.

Change-Id: Ieea461fbc510e966b1c22d0dbb535eed427594eb
Reviewed-by: Orkun Tokdemir <orkun.tokdemir@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>