Add customUiDesigner Usage
* Add `vscode-qt-tools.customUiDesignerExePath` config variable
* Show a warning message when an invalid customUiDesignerExePath is
used

Currently `custom ui designer` overrides the ui designer in the
selected kit. If an invalid paths is given while the custom ui designer
is open, it stays running but it cannot be used to open new `.ui`
files

Note: Since there is no `cmake.onSelectedKitChange()`, the extension
opens the ui designer with the previous kit.

Change-Id: Iabb3b2aff8f29fab88534db5e1f799049170f273
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>