Tweak 'Register Qt Installation' UX
Remove the weirdly capitalized 'The'. Use the term 'directory' and not
just 'path' (which could also be a file path). Remove the message that
is displayed after the user canceled the dialog. The user doesn't need
to be reminded of every action (e.g. don't tell them "Now, you clicked a
button").

Change-Id: I6e4b6e2819767ba80f90f7eac777336512ace632
Reviewed-by: Orkun Tokdemir <orkun.tokdemir@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>