Command: Add documentation commands
* Add the `documentationHomepage` command
This command opens `https://doc.qt.io`

* Add the `documentationSearchManually` command
This commands shows an input box. After the value is given, it checks
whether it is an exact match or not. If it is, it opens the
documentation page directly. If it is not, it shows related results
in a quick pick window.

* Add the `documentationSearchForTheCurrentWord` command
This command searches the word which the cursor is at directly. It
checks whether it is an exact match or not. If it is, it opens the
documentation page directly. If it is not, it shows related results in
a quick pick window.

Change-Id: I985eca969006a8941d4d24df22e2e47bec93c9bc
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>