Kit: Add more information to kit messages
When both global and workspace `qtFolder`s are used, messages can be
confusing for users because while one has installations other might not
have and this can cause showing both `Found` and `Not found` messages
at the same time. This commit adds the location to those messages to
inform better.

Change-Id: I49f8521389aeb0479a1c8338211f8f953d001d7e
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>