Add empty .vscodeignore files to suppress warnings
The below warning is shown when the extension are published to the marketplace.
This commit adds empty .vscodeignore files to suppress warnings.

` WARNING  Neither a .vscodeignore file nor a "files" property in package.json was found. To ensure only necessary files are included in your extension, add a .vscodeignore file or specify the "files" property in package.json. More info: https://aka.ms/vscode-vscodeignore`

Change-Id: I5fa62161034123a18f9f5e596952dbb1b7d5ef80
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>