qt-core: Add qt-core.additionalQtPaths
This new configuration value can be used to specify Qt installations
that were not provided by the Qt installer. The values are absolute
paths to the qtpaths or qmake executables of a Qt installation.

For values in qt-core.additionalQtPaths we now generated CMake kits
like it was already the case for qt-core.qtInstallationRoot.

Fixes: VSCODEEXT-86
Change-Id: Id8eefe92c930546ede2639e18792fc78ba6558dd
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Orkun Tokdemir <orkun.tokdemir@qt.io>