qt-core: Add the `name` property to `additionalQtPaths`
Problem: Sometimes kits with the same name can be generated according to
the query results from either `qtpaths` or `qmake`. This can cause the
user to select the wrong kit.

To avoid this, we need to add a new `name` property to
`additionalQtPaths`. This will allow users to specilaize their kits
with different names. Otherwise, the name will be the same as the
`qtpaths` or `qmake` query result.

* Add new type for `additionalQtPaths`
* Convert the old string parameter to the new type to have uniformity

Change-Id: I9c734c0c69cc19313e66808dad49ac72316f4243
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>