qt-lib: Add deepEqual comparison
The current implementation of the isEqualArrays function compares arrays
by checking if they have the same length and if each corresponding
element is strictly equal (===). This approach works for primitive types
but may fail for complex types like objects or arrays, where strict
equality does not account for deep equality.

To fix this, we use a deep equality check for each element in the array.

Change-Id: I5fae38b205b5a971c1a32dffd385cb489461401f
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>