qt-qml: Add QML Debug Adapter (part 1/x)
This commit introduces the QML Debug Adapter for Visual Studio Code, which
enables debugging of QML applications. The current implementation in QtCreator
was ported to `qt-qml`

Supported features include:

- Breakpoints
- Step over
- Step into
- Step out
- Continue

New launch.json configuration snippets:

- Qt: QML: Attach by port
    It includes basic configuration for attaching to a running QML application.

- Qt: Debug with cppdbg and QML debugger
    It includes basic configuration and command line parameter for QML debugging.

New commands and tasks:

- Qt: Acquire Port
    This task is used with compound debugging configurations to acquire the port
    so that both cppdbg and QML debugger use the same port.
- qt-qml.debugPort
    This command is used pass the same port to both cppdbg and QML debugger
    when using compound debugging configurations.

Some classes were added with this commit:

* QmlEngine

It has the most of the functionality of the QML Debug Adapter. It is responsible
for communicating with the QML debugger and handling debugging events.

* DataStream

This class is used to serialize and deserialize data for the QML Debug Adapter
protocol. It provides methods for writing and reading various data types. It is
a replacement for the `QDataStream` class in Qt.

* QRCParser

It parses .qrc files to extract resource information. It is used to match QML
files shared between the QML Debug Adapter and pysical files.

* QmlEngineUI

This class handles UI elements and interactions for the QML Debug Adapter.

* Timer

This class provides a timer functionality for the QML Debug Adapter. It is used
to schedule tasks and manage timeouts during startup. It is a replacement
for the `QTimer` class in Qt.

Task-number: [VSCODEEXT-14](https://bugreports.qt.io/browse/VSCODEEXT-14)