qt-ui: Fix switching between customWidgetsDesignerExePath and kit paths
When `qt-ui.customWidgetsDesignerExePath` is provided by the user, it
was not used. The below gerrit patch caused this problem. This commit
fixes that.

There was also a bug when `qt-ui.customWidgetsDesignerExePath` is
removed from the settings. The extension was not switching back to the
anymore. This commit fixes that as well.

https://codereview.qt-project.org/c/%7Bgraveyard%7D/qt-labs/vscodeext/+/609633

Fixes: [VSCODEEXT-141](https://bugreports.qt.io/browse/VSCODEEXT-141)
(cherry picked from commit dba2626f3f87c7e3861db0ac0685bde111499451)