qt-lib: Align user home and user local dir with the cmake extension
When we use a hardcoded path for the user home and user local dir, it
causes issues when the user has a different home directory. This change
aligns the user home and user local dir with the cmake extension.

Also, if vscode is installed via a package manager like snap, the cmake
extension generates  `cmake-tools-kits.json` file in the snap directory
, but the Qt extension generates it in the user local dir. This causes
users to not see the kits.

Fixes: [VSCODEEXT-146](https://bugreports.qt.io/browse/VSCODEEXT-146)