Remove our copies of .natvis.xml files
Add the vs-debugtools repo as git submodule.
Change the GitHub workflows to use that submodule.

Add a prepare_natvis_files.ts script that copies the natvis files from
the submodule to qt-cpp/res/natvis and removes the ##NAMESPACE##
markers.

Add a prepareNatvisFiles step to package.json and run this before
compiling qt-cpp.