qt-core/webview: Increase retry duration for /ready check to qtcli
When webview starts, it launches qtcli as a REST server
and waits for it to become ready by polling the /ready endpoint.

This patch increases the retry limit to allow more time for startup,
extending the wait window from 4 to 15 seconds.
It is to make it safer to use in slower environments.

(cherry picked from commit 44248c8973ca576bcc44093fb60f8f00a3f03e31)