qt-cpp: qt-qml: qt-ui: Remove waiting for qt-cpp to be activated
When https://bugreports.qt.io/browse/QTBUG-131702 occurred, as a
workaround, we updated communication between extensions with
https://codereview.qt-project.org/c/%7Bgraveyard%7D/qt-labs/vscodeext/+/609633
and organized the activation order with
https://codereview.qt-project.org/c/%7Bgraveyard%7D/qt-labs/vscodeext/+/609564.
So that qmlls is started only once during the startup. However, we plan
to implement qt-python, and a project will not activate both qt-cpp and
qt-qml due to their nature. The rest of the implementation is being
designed under this condition.

In this situation, if qt-qml activates qt-cpp explicitly, both qt-python
and qt-cpp will be activated for the same project. This breaks our
design. So, we should remove activating qt-cpp from qt-ui and qt-qml
explicitly. If the https://bugreports.qt.io/browse/QTBUG-131702 persists,
it should be fixed on the qmlls side.

* Remove waiting for qt-cpp to be activated in qt-qml and qt-ui
* Send notification via coreAPI during activation of qt-cpp

Fixes: [VSCODE-209](https://bugreports.qt.io/browse/VSCODEEXT-209)