qt-cpp: qt-qml: qt-ui: Obtain qt-core version from package.json of qt-core extension
In the previous implementation, when we bumped the version of qt-core
extension, we were also supposed to update the version in qt-lib.
Instead of doing that, we can directly read the version from qt-core's
package.json. This way, we ensure that qt-lib always uses the correct
version of qt-core without needing to manually update it.