qt-qml: Always restart qmlls when a new version is installed
If we install a new version of qmlls without stopping the currently
running instance, the old version remains active and cannot be replaced.
To fix this, qmlls should always be stopped before installing a new
version.

The root cause of [VSCODEEXT-244](https://bugreports.qt.io/browse/VSCODEEXT-244)
is that [commit 884feba](https://github.com/qt-labs/vscodeext/pull/200/commits/884febaab490d0f33e7e786bbbabf52427ca4ce0)
removed waiting for `qt-cpp` activation and notifying other extensions.
This caused qmlls to start immediately when VS Code starts while the
popup for updating qmlls is still open.

Before that change, when a new qmlls version was published, a popup was
shown to the user and qmlls was not started until the user made a choice.

After the change, qmlls starts immediately, which causes the old version
to keep running and block the update.

This commit fixes the issue by always stopping qmlls before installing
a new version and restarting it before the user makes a choice.

* Improve log messages

Fixes: [VSCODEEXT-242](https://bugreports.qt.io/browse/VSCODEEXT-242)
Fixes: [VSCODEEXT-244](https://bugreports.qt.io/browse/VSCODEEXT-244)