qt-lib, qt-cpp, qt-ui, qt-python: Introduce `QtWorkspaceFeatures`
Previously the `QtWorkspaceType` enumeration was used to indicate
which type of project a specific workspace folder contained.
A folder could only represent one of the available types.

However, there was a need to handle the case where a folder could
contain both a `PySide6` project file and a `CMakeLists.txt` file,
which is logically possible.

This commit addresses that by introducing `QtWorkspaceFeatures`,
which is an object that stores detailed information about a given
workspace folder.

With this change, a workspace folder can be both a CMake project and
a PySide6 project at the same time, instead of being forced to choose
one type exclusively.

This makes multiple extensions such as qt-cpp and qt-python independent.
Each extension can claim that a workspace folder matches the project
type it supports.

Also, this enables the qt-ui extension to locate the Qt Widgets designer
binary across all relevant project types, using it's own priotization
scheme.

The previous `QtWorkspaceType` has been removed accordingly.