qt-ui, qt-python: Support opening Qt Widgets Designer from PySide6
This patch enables qt-ui to locate and launch the Qt Widgets Designer
executable provided within a PySide6 installation.

A new core API key `venvBinPath` is introduced to store the path to
the `bin` (or `Scripts` on Windows) directory of the active virtual
environment, and its value will be updated by `qt-python` whenever
the virtual environment changes.