qt-python: Introduce a new command for PySide6 installation
This commit adds the `qt-python.installPySide6` command.
It first checks the installation status of PySide6 and then shows a
picker for the user to select which version they would like to install.

The first item allows installation from PyPI.
Picker can also include local installation sources under
`<Qt installation root>/QtForPython`.
This UI construction is similar to that of Qt Creator.

If no virtual environment is selected, a popup will appear guiding
the user to create or select a virtual environment first.