qt-core: Improve TS file support
Introduce a new language ID `qt-ts` for Qt translation files.

Whenever a `.ts` file is opened, the extension checks whether it is
a Qt translation file by checking a few lines from the beginning of
the file and searching for the opening <TS> tag.

For the full definition of `qt-ts` language, the following are added:
- Two icons for light and dark themes
- A grammar file that simply includes the `text.xml` pattern.
- A language configuration sourced from VS Code's built-in XML resource.

To further improve usability,
this patch adds a `menus/editor/title` contribution point.
A button for executing the `qt-core.openInLinguist` command will appear
at the top-right corner of the editor tab, only when the file is of
type `qt-ts`.

Task-number: [VSCODEEXT-25](https://bugreports.qt.io/browse/VSCODEEXT-25)