qt-python: Fix picking up global environment when running command
The previous implementation produced an empty string for a virtual
environment activation command, which was not considered as an error
by the shell.

In such cases, the following command would succeed, but it would run
in the global environment, causing unintended behavior.

For example, when retrieving PySide6 installation information by running
a `pip` command with no virtual environment installed, it would fetch
the information from the global environment which is incorrect.

This patch fixes the issue.